refactor(connector-sdk): drop connector-specific identity from core
Remove the SDK IDENTITY constants and GitHub/Google normalizers from the
shared ingestion path. Connectors own their namespace strings and
pre-normalize values; the SDK dispatcher keeps only generic phone/email/
auth_user_id hygiene. GitHub ACL graph code gets local normalizers.
