+++ examples/brand-intelligence/__tests__/connector-sdk.mock.ts
@@ -125,18 +125,6 @@ export function connectorSdkMock() {
             : ((checkpoint?.last_timestamp as string | undefined) ?? null),
       },
     }),
-    IDENTITY: {
-      PHONE: "phone",
-      EMAIL: "email",
-      WA_JID: "wa_jid",
-      SLACK_USER_ID: "slack_user_id",
-      GITHUB_LOGIN: "github_login",
-      GITHUB_USER_ID: "github_user_id",
-      GITHUB_REPO_ID: "github_repo_id",
-      GITHUB_REPO_FULL_NAME: "github_repo_full_name",
-      AUTH_USER_ID: "auth_user_id",
-      GOOGLE_CONTACT_ID: "google_contact_id",
-    },
     extensionDomScrape: async (opts: DomScrapeOpts) => {
       const observation = await opts.dispatcher.dispatch("navigate", {
         cs_scrape: true,
