+++ packages/server/src/__tests__/integration/authz/github-repo-visibility.test.ts
@@ -14,7 +14,7 @@
  * events/search-content-visibility.test.ts).
  */
 
-import { normalizeGithubRepoFullName } from '@lobu/connector-sdk';
+import { normalizeGithubRepoFullName } from '../../../authz/github-normalize.js';
 import { beforeAll, beforeEach, describe, expect, it } from 'vitest';
 import { syncGithubConnectionAcl } from '../../../authz/github-acl-sync';
 import { buildGithubRepoGraph } from '../../../authz/github-repo-graph';
