+++ packages/server/src/__tests__/integration/events/resource-gate-stale-acl.test.ts
@@ -17,7 +17,7 @@
  * deterministic without a live embedding service (mirrors github-repo-visibility).
  */
 
-import { normalizeGithubRepoFullName } from '@lobu/connector-sdk';
+import { normalizeGithubRepoFullName } from '../../../authz/github-normalize.js';
 import { beforeAll, beforeEach, describe, expect, it } from 'vitest';
 import { buildGithubRepoGraph } from '../../../authz/github-repo-graph';
 import type { ToolContext } from '../../../tools/registry';
